dt-learninghub/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.9
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==36.0.1
    # via
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==2.2.25
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   drf-nested-routers
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-rbac
django-cors-headers==3.10.1
    # via -r requirements/base.in
django-crum==0.7.9
    # via
    #   edx-django-utils
    #   edx-rbac
django-extensions==3.1.5
    # via -r requirements/base.in
django-model-utils==4.2.0
    # via
    #   -r requirements/base.in
    #   edx-rbac
django-waffle==2.2.1
    # via
    #   -r requirements/base.in
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.13.1
    # via
    #   -r requirements/base.in
    #   drf-jwt
    #   drf-nested-routers
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-drf-extensions
drf-jwt==1.19.1
    # via edx-drf-extensions
drf-nested-routers==0.93.4
    # via -r requirements/base.in
drf-yasg==1.20.0
    # via edx-api-doc-tools
edx-api-doc-tools==1.5.0
    # via -r requirements/base.in
edx-auth-backends==4.0.1
    # via -r requirements/base.in
edx-django-release-util==1.1.0
    # via -r requirements/base.in
edx-django-utils==4.4.1
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==8.0.1
    # via
    #   -r requirements/base.in
    #   edx-rbac
edx-opaque-keys==2.2.2
    # via edx-drf-extensions
edx-rbac==1.5.1
    # via -r requirements/base.in
edx-rest-api-client==5.4.0
    # via -r requirements/base.in
future==0.18.2
    # via pyjwkest
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.0.1
    # via jinja2
mysqlclient==2.1.0
    # via -r requirements/base.in
newrelic==7.2.4.171
    # via edx-django-utils
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via drf-yasg
pbr==5.8.0
    # via stevedore
psutil==5.8.0
    # via edx-django-utils
pycparser==2.21
    # via cffi
pycryptodomex==3.12.0
    # via pyjwkest
pyjwkest==1.4.2
    # via edx-drf-extensions
pyjwt[crypto]==2.3.0
    # via
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pymongo==4.0.1
    # via edx-opaque-keys
pyparsing==3.0.6
    # via packaging
python-dateutil==2.8.2
    # via edx-drf-extensions
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.3
    # via
    #   -r requirements/base.in
    #   django
    #   djangorestframework
pyyaml==6.0
    # via edx-django-release-util
requests==2.26.0
    # via
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
ruamel.yaml==0.17.17
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
rules==3.0
    # via -r requirements/base.in
semantic-version==2.8.5
    # via edx-drf-extensions
six==1.16.0
    # via
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   edx-rbac
    #   pyjwkest
    #   python-dateutil
slumber==0.7.1
    # via edx-rest-api-client
social-auth-app-django==5.0.0
    # via edx-auth-backends
social-auth-core==4.1.0
    # via
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.2
    # via django
stevedore==3.5.0
    # via
    #   edx-django-utils
    #   edx-opaque-keys
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via requests
